backend/requirements.txt
reportlab>=4.0.0
requests>=2.28.0
python-dateutil>=2.8.0
fastapi>=0.115.0
uvicorn>=0.30.0
PyJWT>=2.8.0
bcrypt>=4.1.0
psycopg2-binary>=2.9.0
